Bug 1605134 - Break out WaitFor helpers to a dedicated file. r=padenot
This allows re-using them in TestAudioCallbackDriver.cpp.

Differential Revision: https://phabricator.services.mozilla.com/D97405
